requirements.txt
altgraph==0.17.2
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
Django==4.1.6
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-fontawesome-5==1.0.18
django-js-asset==2.1.0
django-matplotlib==0.1
django-model-utils==4.3.1
fonttools==4.42.0
future==0.18.2
html5lib==1.1
idna==3.4
importlib-resources==6.0.1
kiwisolver==1.4.4
lxml==4.9.3
macholib==1.15.2
matplotlib==3.7.2
numpy==1.25.2
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pydotplus==2.0.2
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyparsing==3.0.9
pypdf==3.15.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.15.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
treelib==1.7.0
typing_extensions==4.7.1
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
webencodings==0.5.1
xhtml2pdf==0.2.11
zipp==3.16.2
